x264: update to r2197 and fix clang optimization
Upgrade x264 to the latest stable tarball, 20120425.
Add `ENV.O1` when compiling with clang, which fixes a segfault
when encoding with libx264.  The fps encoding rate remains the
same using this commit.  The changelog listed in the formula
comments was checked, and the version is unchanged.  This will
cause a hash fail when x264-r2184.tar.bz2 already exists.  The
user will have to remove the cached tarball and try again.  At
the time this commit is being submitted, the x264-devel mailing
listserv is offline.  So it hasn't been reported a second time.

Fixes #11248.
Closes #11885.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>