rasqal 0.9.29
Upgrade raptor to version 0.9.29 and adjust the location of the
html docs from gtk-doc to the usual location in share.  For this
app, the standard helper path `doc` results in `doc/rasqal/rasqal`.
So `share/doc` is used to eliminate the duplicate `rasqal` path.

Note: The homepage for rasqal mentions an upcoming API change:
'RDQL support will be removed in the next release (0.9.30)'

Closes #12259.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>