ncmpcpp: add pkg-config, update fails_with
ncmpcpp has a build-time dep on pkg-config so that it can find
libmpdclient, and it needs the `fails_with` version updated
because it still has problems with clang.

Add a dep on pkg-config
Update the fails_with build.
Use the new options DSL.
Tested on ML with clang and llvm from XCode-4.4.1.

Fixes #14160

Closes #14172.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>